Moved the Matrix and Vector classes to their own namespace.
